do not start gnirehtet as ./

+++ woof.sh
@@ -11,15 +11,27 @@
 # - after gnirehtet restarts, the phone connects again automatically, and
 #   the reverse tethering is restored
 
+which gnirehtet 2>/dev/null
+if [ $? -ne 0 ]; then
+    PATH="$PATH:."
+fi
+which gnirehtet 2>/dev/null
+if [ $? -ne 0 ]; then
+    echo "gnirehtet not found"
+    echo "either install it through a package for your OS"
+    echo "or download the Rust binary release from GitHub"
+    exit -1
+fi
+
 while true; do
-    echo "(re)starting ./gnirehtet"
-    ./gnirehtet autorun | {
+    echo "(re)starting gnirehtet"
+    gnirehtet autorun | {
         while IFS= read -r line
         do
             echo "$line"
             if [[ "$line" =~ "disconnected" ]]; then
                 kill $(pgrep -f gnirehtet)
-                echo "./gnirehtet killed"
+                echo "gnirehtet killed"
                 sleep 1
             fi
         done
